Feature/unreal windows plugin (OCS-442)

See merge request one/ardentblue/one-game-sdk!179

+++ one/arcus/internal/time.cpp
@@ -3,7 +3,7 @@
 namespace i3d {
 namespace one {
 
-IntervalTimer::IntervalTimer(unsigned int seconds)
+IntervalTimer::IntervalTimer(long long seconds)
     : _interval(seconds), _last_trigger_time(steady_clock::duration::zero()) {}
 
 bool IntervalTimer::update() {
